+++ test/integration/src/segments/user/UserSegment.js
@@ -2,15 +2,11 @@ import MapSegment from 'soya/lib/data/redux/segment/map/MapSegment.js';
 import UserActionCreator from './UserActionCreator.js';
 
 export default class UserSegment extends MapSegment {
-  static get segmentName() {
+  static id() {
     return 'user';
   }
 
-  _getName() {
-    return UserSegment.segmentName;
-  }
-
   _createActionCreator() {
-    return new UserActionCreator(this._getName());
+    return new UserActionCreator(UserSegment.id());
   }
 }